fix(07-02): correct compilation errors in MarketPrediction module
- MlServiceConfig: add explicit (int) cast for Duration.toSeconds() long-to-int
- MarketPredictionController: fix @RateLimit attributes (limit/period instead of permits/seconds)
- MarketPredictionServiceTest: use setId() instead of builder .id() for BaseEntity
- MarketPredictionControllerTest: fix MockBean import (mock.mockito instead of mock.bean)

+++ oaiss-chain-backend/src/main/java/com/oaiss/chain/config/MlServiceConfig.java
@@ -37,8 +37,8 @@ public class MlServiceConfig {
                         (int) connectTimeout.toMillis())
                 .responseTimeout(readTimeout)
                 .doOnConnected(conn -> conn
-                        .addHandlerLast(new ReadTimeoutHandler(readTimeout.toSeconds()))
-                        .addHandlerLast(new WriteTimeoutHandler(readTimeout.toSeconds())));
+                        .addHandlerLast(new ReadTimeoutHandler((int) readTimeout.toSeconds()))
+                        .addHandlerLast(new WriteTimeoutHandler((int) readTimeout.toSeconds())));
 
         return webClientBuilder
                 .baseUrl(url)
